Release candidate

+++ ui/lib/src/elements/event-detail.ts
@@ -20,13 +20,15 @@ import '@shoelace-style/shoelace/dist/components/alert/alert.js';
 import '@shoelace-style/shoelace/dist/components/icon-button/icon-button.js';
 import '@shoelace-style/shoelace/dist/components/button/button.js';
 import '@shoelace-style/shoelace/dist/components/icon/icon.js';
+import '@shoelace-style/shoelace/dist/components/relative-time/relative-time.js';
+import '@shoelace-style/shoelace/dist/components/tag/tag.js';
 import '@shoelace-style/shoelace/dist/components/spinner/spinner.js';
 import '@shoelace-style/shoelace/dist/components/skeleton/skeleton.js';
 
 import '@darksoil/assemble/dist/elements/call-to-action-needs.js';
 import '@darksoil/assemble/dist/elements/call-to-action-need-progress.js';
 
-import './attendees-for-event.js';
+import './participants-for-event.js';
 import './edit-event.js';
 
 import { gatherStoreContext } from '../context.js';
@@ -34,11 +36,16 @@ import { GatherStore } from '../gather-store.js';
 import { Event } from '../types.js';
 import {
   mdiCalendarClock,
+  mdiCancel,
   mdiCash,
+  mdiCheckDecagram,
   mdiDelete,
   mdiMapMarker,
   mdiPencil,
 } from '@mdi/js';
+import { SlDialog } from '@shoelace-style/shoelace';
+import { CallToAction } from '@darksoil/assemble';
+import { isExpired } from '../utils.js';
 
 @localized()
 @customElement('event-detail')
@@ -55,15 +62,17 @@ export class EventDetail extends LitElement {
   /**
    * @internal
    */
-  _event = new StoreSubscriber(this, () =>
-    this.gatherStore.events.get(this.eventHash)
+  _event = new StoreSubscriber(
+    this,
+    () => this.gatherStore.eventsCallToActionsAndAssemblies.get(this.eventHash),
+    () => [this.eventHash]
   );
 
   /**
    * @internal
    */
-  _attendees = new StoreSubscriber(this, () =>
-    this.gatherStore.attendeesForEvent.get(this.eventHash)
+  _participants = new StoreSubscriber(this, () =>
+    this.gatherStore.participantsForEvent.get(this.eventHash)
   );
 
   /**
@@ -72,12 +81,24 @@ export class EventDetail extends LitElement {
   @state()
   _editing = false;
 
+  /**
+   * @internal
+   */
+  @state()
+  _cancelling = false;
+
+  /**
+   * @internal
+   */
+  @state()
+  _approving = false;
+
   async deleteEvent(event: EntryRecord<Event>) {
+    if (this._cancelling) return;
+
+    this._cancelling = true;
     try {
       await this.gatherStore.client.deleteEvent(this.eventHash);
-      await this.gatherStore.assembleStore.client.closeCallToAction(
-        event.entry.call_to_action_hash
-      );
 
       this.dispatchEvent(
         new CustomEvent('event-deleted', {
@@ -88,121 +109,236 @@ export class EventDetail extends LitElement {
           },
         })
       );
+
+      (
+        this.shadowRoot?.getElementById('cancel-event-dialog') as SlDialog
+      ).hide();
     } catch (e: any) {
       notifyError(msg('Error deleting the event'));
       console.error(e);
     }
+    this._cancelling = false;
   }
 
-  renderDetail(entryRecord: EntryRecord<Event>) {
-    const amIAuthor =
-      entryRecord.action.author.toString() ===
-      this.gatherStore.client.client.myPubKey.toString();
+  async approveProposal(event: EntryRecord<Event>) {
+    if (this._approving) return;
+
+    this._approving = true;
+    try {
+      await this.gatherStore.assembleStore.client.createAssembly({
+        call_to_action_hash: event.entry.call_to_action_hash,
+        satisfactions_hashes: [],
+      });
+
+      (
+        this.shadowRoot?.getElementById('approve-proposal-dialog') as SlDialog
+      ).hide();
+    } catch (e: any) {
+      notifyError(msg('Error approving the proposal'));
+      console.error(e);
+    }
+    this._cancelling = false;
+  }
+
+  renderApproveProposalDialog(entryRecord: EntryRecord<Event>) {
     return html`
-      <sl-card class="column" style="width: 700px">
+      <sl-dialog id="approve-proposal-dialog" .label=${msg('Approve Proposal')}>
+        <span
+          >${msg(
+            'Are you sure you want to approve this proposal and convert it to an event?'
+          )}</span
+        >
+
+        <sl-button
+          slot="footer"
+          variant="primary"
+          .loading=${this._approving}
+          @click=${() => this.approveProposal(entryRecord)}
+          >${msg('Approve Proposal')}</sl-button
+        >
+      </sl-dialog>
+    `;
+  }
+
+  renderCancelEventDialog(entryRecord: EntryRecord<Event>) {
+    return html`
+      <sl-dialog id="cancel-event-dialog" .label=${msg('Cancel Event')}>
+        <span>${msg('Are you sure you want to cancel this event?')}</span>
+
+        <sl-button
+          slot="footer"
+          variant="danger"
+          .loading=${this._cancelling}
+          @click=${() => this.deleteEvent(entryRecord)}
+          >${msg('Cancel Event')}</sl-button
+        >
+      </sl-dialog>
+    `;
+  }
+
+  renderStatus(
+    isCancelled: boolean,
+    callToAction: EntryRecord<CallToAction>,
+    assemblies: ActionHash[]
+  ) {
+    if (isCancelled)
+      return html`<sl-tag variant="warning">${msg('Cancelled')}</sl-tag>`;
+    if (isExpired(callToAction.entry) && assemblies.length === 0)
+      return html`<sl-tag variant="warning">${msg('Expired')}</sl-tag>`;
+    if (assemblies.length > 0)
+      return html`<sl-tag variant="success">${msg('Event')}</sl-tag>`;
+
+    return html`<sl-tag
+      >${msg('Proposal')}${callToAction.entry.expiration_time
+        ? html`${msg(': expires')}&nbsp;
+            <sl-relative-time
+              .date=${callToAction.entry.expiration_time / 1000}
+            ></sl-relative-time>`
+        : html``}</sl-tag
+    >`;
+  }
+
+  renderActions(
+    event: EntryRecord<Event>,
+    isCancelled: boolean,
+    callToAction: EntryRecord<CallToAction>,
+    assemblies: ActionHash[]
+  ) {
+    const amIAuthor =
+      event.action.author.toString() ===
+      this.gatherStore.client.client.myPubKey.toString();
+    if (
+      isCancelled ||
+      (isExpired(callToAction.entry) && assemblies.length === 0) ||
+      !amIAuthor
+    )
+      return html``;
+
+    return html`
+      <sl-icon-button
+        title=${msg('Edit event')}
+        @click=${() => {
+          this._editing = true;
+        }}
+        .src=${wrapPathInSvg(mdiPencil)}
+      ></sl-icon-button>
+      <sl-icon-button
+        title=${msg('Cancel event')}
+        style="margin-left: 8px"
+        .src=${wrapPathInSvg(mdiCancel)}
+        @click=${() =>
+          (
+            this.shadowRoot?.getElementById('cancel-event-dialog') as SlDialog
+          ).show()}
+      ></sl-icon-button>
+      ${assemblies.length === 0
+        ? html`
+            <sl-icon-button
+              title=${msg('Approve proposal')}
+              style="margin-left: 8px"
+              .src=${wrapPathInSvg(mdiCheckDecagram)}
+              @click=${() =>
+                (
+                  this.shadowRoot?.getElementById(
+                    'approve-proposal-dialog'
+                  ) as SlDialog
+                ).show()}
+            ></sl-icon-button>
+          `
+        : html``}
+    `;
+  }
+
+  renderDetail(
+    event: EntryRecord<Event>,
+    isCancelled: boolean,
+    callToAction: EntryRecord<CallToAction>,
+    assemblies: ActionHash[]
+  ) {
+    return html`
+      ${this.renderApproveProposalDialog(event)}
+      ${this.renderCancelEventDialog(event)}
+      <sl-card class="column">
         <show-image
           slot="image"
-          .imageHash=${entryRecord.entry.image}
+          .imageHash=${event.entry.image}
           style="width: 700px; height: 300px; flex-basis: 0;"
         ></show-image>
 
-        <div style="display: flex; flex-direction: column; flex: 1;">
-          <div style="display: flex; flex-direction: row">
-            <span
-              class="title"
-              style="flex: 1; margin-top: 16px; margin-bottom: 16px;"
-              >${entryRecord.entry.title}</span
+        <div class="row" style="flex: 1">
+          <div style="display: flex; flex-direction: column; flex: 1;">
+            <span class="title" style="margin-bottom: 16px"
+              >${event.entry.title}</span
             >
 
-            ${amIAuthor
-              ? html`
-                  <sl-icon-button
-                    title=${msg('edit')}
-                    style="margin-left: 8px"
-                    @click=${() => {
-                      this._editing = true;
-                    }}
-                    .src=${wrapPathInSvg(mdiPencil)}
-                  ></sl-icon-button>
-                  <sl-icon-button
-                    title=${msg('delete event')}
-                    style="margin-left: 8px"
-                    .src=${wrapPathInSvg(mdiDelete)}
-                    @click=${() => this.deleteEvent(entryRecord)}
-                  ></sl-icon-button>
-                `
-              : html``}
-            <slot name="action"></slot>
+            <span style="white-space: pre-line; margin-bottom: 16px"
+              >${event.entry.description}</span
+            >
+
+            <div class="column" style="justify-content: end; flex: 1">
+              <div
+                style="display: flex; flex-direction: row; align-items: center;"
+              >
+                <sl-icon
+                  title=${msg('location')}
+                  style="margin-right: 4px"
+                  .src=${wrapPathInSvg(mdiMapMarker)}
+                ></sl-icon>
+                <span style="white-space: pre-line"
+                  >${event.entry.location}</span
+                >
+              </div>
+
+              <div
+                style="display: flex; flex-direction: row; align-items: center; margin-top: 16px"
+              >
+                <sl-icon
+                  title=${msg('time')}
+                  style="margin-right: 4px"
+                  .src=${wrapPathInSvg(mdiCalendarClock)}
+                ></sl-icon>
+                <span
+                  >${new Date(event.entry.start_time / 1000).toLocaleString()} -
+                  ${new Date(
+                    event.entry.end_time / 1000
+                  ).toLocaleString()}</span
+                >
+              </div>
+
+              ${event.entry.cost
+                ? html` <div
+                    style="display: flex; flex-direction: row; align-items: center; margin-top: 16px"
+                  >
+                    <sl-icon
+                      title=${msg('cost')}
+                      style="margin-right: 4px"
+                      .src=${wrapPathInSvg(mdiCash)}
+                    ></sl-icon>
+                    <span style="white-space: pre-line"
+                      >${event.entry.cost}</span
+                    >
+                  </div>`
+                : html``}
+            </div>
           </div>
 
-          <div style="display: flex; flex-direction: column;">
-            <span style="white-space: pre-line; margin-bottom: 16px;"
-              >${entryRecord.entry.description}</span
+          <div class="column" style="align-items: end">
+            ${this.renderStatus(isCancelled, callToAction, assemblies)}
+            <div
+              class="row"
+              style="justify-content:end; flex: 1; margin-top: 8px"
             >
-
-            <div style="display: flex; flex-direction: row;">
-              <div class="column" style="justify-content: end">
-                <div
-                  title=${msg('location')}
-                  style="display: flex; flex-direction: row; align-items: center;"
-                >
-                  <sl-icon
-                    style="margin-right: 4px"
-                    .src=${wrapPathInSvg(mdiMapMarker)}
-                  ></sl-icon>
-                  <span style="white-space: pre-line"
-                    >${entryRecord.entry.location}</span
-                  >
-                </div>
-
-                <div
-                  title=${msg('time')}
-                  style="display: flex; flex-direction: row; align-items: center"
-                >
-                  <sl-icon
-                    style="margin-right: 4px"
-                    .src=${wrapPathInSvg(mdiCalendarClock)}
-                  ></sl-icon>
-                  <span style="white-space: pre-line"
-                    >${new Date(
-                      entryRecord.entry.start_time / 1000
-                    ).toLocaleString()}
-                    -
-                    ${new Date(
-                      entryRecord.entry.end_time / 1000
-                    ).toLocaleString()}</span
-                  >
-                </div>
-
-                ${entryRecord.entry.cost
-                  ? html` <div
-                      title=${msg('cost')}
-                      style="display: flex; flex-direction: row; align-items: center"
-                    >
-                      <sl-icon
-                        style="margin-right: 4px"
-                        .src=${wrapPathInSvg(mdiCash)}
-                      ></sl-icon>
-                      <span style="white-space: pre-line"
-                        >${entryRecord.entry.cost}</span
-                      >
-                    </div>`
-                  : html``}
-              </div>
-
-              <span style="flex: 1"></span>
-
-              <div class="column" style="justify-content: end">
-                <div
-                  class="row"
-                  style="align-items: center; margin-bottom: 8px;"
-                >
-                  <span style="margin-right: 8px">${msg('Hosted by')}</span>
-                  <agent-avatar
-                    .agentPubKey=${entryRecord.action.author}
-                  ></agent-avatar>
-                </div>
-              </div>
+              ${this.renderActions(
+                event,
+                isCancelled,
+                callToAction,
+                assemblies
+              )}
+              <slot name="action"></slot>
+            </div>
+            <div class="row" style="align-items: center;">
+              <span style="margin-right: 8px">${msg('Hosted by')}</span>
+              <agent-avatar .agentPubKey=${event.action.author}></agent-avatar>
             </div>
           </div>
         </div>
@@ -210,14 +346,16 @@ export class EventDetail extends LitElement {
     `;
   }
 
-  renderEvent(maybeEntryRecord: EntryRecord<Event> | undefined) {
-    if (!maybeEntryRecord)
-      return html`<span>${msg("The requested event doesn't exist")}</span>`;
-
+  renderEvent(
+    event: EntryRecord<Event>,
+    isCancelled: boolean,
+    callToAction: EntryRecord<CallToAction>,
+    assemblies: ActionHash[]
+  ) {
     if (this._editing) {
       return html`<edit-event
         .originalEventHash=${this.eventHash}
-        .currentRecord=${maybeEntryRecord}
+        .currentRecord=${event}
         @event-updated=${async () => {
           this._editing = false;
         }}
@@ -230,18 +368,18 @@ export class EventDetail extends LitElement {
 
     return html`<div class="row" style="justify-content: center">
       <div class="column" style="margin-right: 16px;">
-        ${this.renderDetail(maybeEntryRecord)}
+        ${this.renderDetail(event, isCancelled, callToAction, assemblies)}
 
-        <call-to-action-need
-          .callToActionHash=${maybeEntryRecord.entry.call_to_action_hash}
+        <call-to-action-needs
+          .callToActionHash=${event.entry.call_to_action_hash}
           .hideNeeds=${[0]}
-        ></call-to-action-need>
+        ></call-to-action-needs>
       </div>
       <div class="column">
-        <attendees-for-event
+        <participants-for-event
           style="margin-bottom: 16px;"
           .eventHash=${this.eventHash}
-        ></attendees-for-event>
+        ></participants-for-event>
         <slot name="attachments"></slot>
       </div>
     </div> `;
@@ -256,7 +394,20 @@ export class EventDetail extends LitElement {
           <sl-spinner style="font-size: 2rem"></sl-spinner>
         </div>`;
       case 'complete':
-        return this.renderEvent(this._event.value.value);
+        const event = this._event.value.value[0];
+        const maybeCallToAction = this._event.value.value[1];
+
+        if (!event || !maybeCallToAction || !maybeCallToAction[0])
+          return html`<span
+            >${msg('The requested event was not found.')}</span
+          >`;
+
+        return this.renderEvent(
+          event.record,
+          event.isCancelled,
+          maybeCallToAction[0],
+          maybeCallToAction[1]
+        );
       case 'error':
         return html`<display-error
           .error=${this._event.value.error.data.data}
